+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Automate Squatters event consumption increase
 * Allow dragging actor images from a recipe to another recipe
 * Allow users to re-enable the first run message
+* Add all structure traits to Structure Browser filters
+* Add Gain Fame button if built building type matches kingdom
 
 ### Changed
 * Require FoundryVTT 13